+++ src/components/pages/posts/Comments/ReplyItem.tsx
@@ -1,6 +1,6 @@
-import { useGetUserProfile } from "@/hooks/User/useGetUserProfile";
 import { useDeleteReply } from "@/lib/comment/hooks/useDeleteReply";
 import { ReplyDto } from "@/lib/comment/type";
+import { useGetUserProfile } from "@/lib/user/hooks/useGetUserProfile";
 import { DEFAULT_PROFILE_IMG_CAT } from "@/shared/const/UserprofileImgPath";
 import { useAuthStore } from "@/stores/auth/useAuthStore";
 import { formatTimestamp } from "@/utils/formatTimestamp";
