+++ src/components/pages/user/userList/UserCard.tsx
@@ -1,4 +1,4 @@
-import { useGetUserProfile } from "@/hooks/User/useGetUserProfile";
+import { useGetUserProfile } from "@/lib/user/hooks/useGetUserProfile";
 import { DEFAULT_PROFILE_IMG_CAT } from "@/shared/const/UserprofileImgPath";
 import { useModalStore } from "@/stores/modal/useModalStore";
 import { useNavigate } from "react-router-dom";
