+++ src/lib/user/hooks/useCreateUserProfile.tsx
@@ -2,6 +2,7 @@ import { createUserProfile } from "@/lib/user/api";
 import { UserProfile } from "@/lib/user/type";
 
 import { useMutation, useQueryClient } from "@tanstack/react-query";
+import { USER_PROFILE } from "../key";
 
 export function useCreateUserProfile(userId: string) {
   const queryClient = useQueryClient();
@@ -12,7 +13,7 @@ export function useCreateUserProfile(userId: string) {
     onSuccess: () =>
       queryClient.invalidateQueries({
         exact: true,
-        queryKey: ["userProfile", userId],
+        queryKey: [USER_PROFILE, userId],
       }),
   });
 }
