+++ src/lib/user/hooks/useUpdateUserProfile.tsx
@@ -1,6 +1,7 @@
 import { updateUserProfile } from "@/lib/user/api";
 import { UserProfile } from "@/lib/user/type";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
+import { USER_PROFILE } from "../key";
 
 export function useUpdateUserProfile(userId: string) {
   const queryClient = useQueryClient();
@@ -11,7 +12,7 @@ export function useUpdateUserProfile(userId: string) {
     onSuccess: () =>
       queryClient.invalidateQueries({
         exact: true,
-        queryKey: ["userProfile", userId],
+        queryKey: [USER_PROFILE, userId],
       }),
   });
 }
