+++ src/lib/user/key.ts
@@ -1,0 +1,1 @@
+export const USER_PROFILE = "USER_PROFILE";
